factura/serializers/productSerializer.py
from factura.models.product import Product
from rest_framework import serializers

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id_product', 'product_name', 'product_stock', 'product_price', 'isActive']